// src/app/feature/specifications/models/specificationOriginTypeEnum.ts
export enum SpecificationOriginTypeEnum {
    NotSeted = 0, /// Не установлен
    EstmcBased = 1,/// Спецификация, основанная на ЕСТМЦ
    TruBased = 2, /// Спецификация, основанная на ТРУ
    PurchasePlanBased = 3, /// Спецификация, основанная на ГПЗ
    LotSAPBased = 4, /// Спецификация основанная на Лот (SAP)
}