// src/app/feature/specifications/models/specificationsEntity.ts
export interface SpecificationsEntity {
        ID?: number;
        NUM_SPECIFIC?: string;
        ACCOUNTS_ID?: string;
        CREATE_DATE?: Date;
        MAIN_CATEGORY?: number;
        MAIN_CATEGORY_TRU_ID?: number;
        SPECIFICATION_TYPE?: number;
        REF_CURRENCY_ID?: number;
        CUSTOMERS?: string;
        TOTAL?: number;
        dTOTAL?: number;
        STATUS?: number;
        ISDELETED?: number;
        ISANONS?: number;
        LOT_NUM_SAP?: number;
        CPURCH_REQ_STATUS?: number;
        ORIGIN_TYPE?: number;
        ORIGIN_TYPE_TEXT?: string;
        PURCHASE_TYPE?: number;
        HAS_BALANCE_GPZ?: number;
        STATUS_FORMED_DATE?: Date;
        BLN_MAX_AGE?: number;
        HAS_BALANCE?: number;
        REQUEST_CHECK_OWN_PRODUCTS?: number;
        HAS_OWN_PRODUCTS?: number;
        PRE_PURCHASE_MANNER_ID?: number;
        PURCHASE_MANNER_ID?: number;
        IS_PURCHASE_MANNER_FROM_SAP?: number;
        IS_EMERGENCY_PURCHASE?: number;
        ANALOG_NAME?: string;
        PARENT_SPECIFICATION_ID?: number;
        IS_FRAME_AGREEMENT?: string;
        WSS_DOCS_CARD_ID?: string;
        IsFullyReuseQuoteQuery?: boolean;
        LastModification?: Date;
        STATUS_TEXT?: string;
}