// src/app/feature/specifications/models/specificationStatusEnum.ts
export enum SpecificationStatusEnum{
        Forming = 0,// Формируется (0)
        Formed = 1,// Сформирована (1)
        Lot = 2,  /// Лот (2)
        IsNotReality = 3,// Конкурс/закуп не состоялся (3)
        IsDeletedOnETP = 4 // Объявление удалено на ЭТП (4)
}